Gathered all determinization-related data in a struct to avoid passing many parameters.

+++ re2c/src/dfa/dfa.h
@@ -47,7 +47,7 @@ struct dfa_state_t
 
 struct dfa_t
 {
-	static const size_t NIL;
+	static const uint32_t NIL;
 
 	std::vector<dfa_state_t*> states;
 	const size_t nchars;
